python-api/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Analyse de données
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2
xlrd==2.0.1

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4

# Base de données
pymongo==4.6.0
motor==3.3.2

# Utilitaires
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2

# Analyse de fichiers
chardet==5.2.0
python-magic==0.4.27

# Logs et monitoring
loguru==0.7.2

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1